+++ utils/handlers/EntryInput.ts
@@ -1,9 +1,8 @@
 import { certainKeyPressed } from 'lunchbox';
 import { useState } from 'preact/hooks';
 import { useTagList } from 'hooks';
-import { setEntry } from 'db/middleware.ts';
 import { isURL } from 'utils';
-import { iEntryInput } from '../../islands/EntryInput/index.tsx';
+import { iEntryInput } from 'islands/EntryInput/index.tsx';
 
 type Steps = 'entrymark' | 'tags';
 
@@ -17,6 +16,36 @@ export default function (props: iEntryInput) {
     inputStep,
     setInputStep,
   ] = useState<(Steps)[]>([]);
+
+  async function setEntry(_ev: KeyboardEvent) {
+    const setApiUrl = entry._id
+      ? `/api/entries/${entry._id}/update`
+      : '/api/entries/create';
+
+    await fetch(setApiUrl, {
+      method: 'POST',
+      mode: 'no-cors',
+      body: JSON.stringify({
+        utc_created_at: entry.utc_created_at,
+        content: entryValue,
+        tags: tags,
+        entry_mark: entryMark,
+      }),
+    })
+      .then(() => {
+        if (updateEntriesSignal) {
+          updateEntriesSignal.value++;
+        }
+      })
+      .catch((e) => {
+        alert('Create entry error.');
+        console.error('Create entry error:', e);
+      });
+    setEntryValue('');
+    setEntryMark('');
+    updateTags([], tags);
+    setInputStep([]);
+  }
 
   function handleEntryInput(ev: Event) {
     const value = (ev.target as HTMLTextAreaElement).value;
@@ -86,20 +115,6 @@ export default function (props: iEntryInput) {
     }
   };
 
-  const handleCreateEntryShortcut = (_ev: KeyboardEvent) => {
-    setEntry({
-      id: entry.id,
-      created_at: entry.created_at,
-      content: entryValue,
-      tags: tags,
-      entry_mark: entryMark,
-    });
-    setEntryValue('');
-    setEntryMark('');
-    updateTags([], tags);
-    setInputStep([]);
-  };
-
   const handleEntryMarkInput = (ev: KeyboardEvent) => {
     const entry_mark = (ev.target as HTMLInputElement).value;
     const isMarkUrl = isURL(entry_mark);
@@ -115,12 +130,9 @@ export default function (props: iEntryInput) {
 
   const handleConatinerKeyDown = (ev: KeyboardEvent) => {
     if ((ev.metaKey || ev.ctrlKey) && ev.key === 'Enter') {
-      handleCreateEntryShortcut(ev);
+      setEntry(ev);
       if (onFocusOut) {
         onFocusOut();
-      }
-      if (updateEntriesSignal) {
-        updateEntriesSignal.value++;
       }
     }
     if (ev.key === 'Escape' && onFocusOut) {