+++ utils/types.ts
@@ -1,3 +1,7 @@
 export type dbItem<T, idT = number> = {
-  id: idT;
+  _id: idT;
+} & T;
+
+export type docItem<T> = {
+  [key: string]: unknown;
 } & T;